Refacto directory organisation, remove application translation (delete language file). Add resource files (maps). Simplify .gitignore
