--- main.cpp
@@ -1,24 +1,0 @@
-#include "mainwindow.hpp"
-#include <iostream>
-#include <QApplication>
-#include <QLocale>
-#include <QTranslator>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-
-    std::cout << "Hello cutie!" << std::endl;
-    QTranslator translator;
-    const QStringList uiLanguages = QLocale::system().uiLanguages();
-    for (const QString &locale : uiLanguages) {
-        const QString baseName = "mod1_" + QLocale(locale).name();
-        if (translator.load(":/i18n/" + baseName)) {
-            a.installTranslator(&translator);
-            break;
-        }
-    }
-    MainWindow w;
-    w.show();
-    return a.exec();
-}
