+++ resources/demo2.mod1
@@ -1,0 +1,2 @@
+(10000,10000,6000)
+(15000,10000,100) (15000,15000,4000)
