+++ resources/demo3.mod1
@@ -1,0 +1,3 @@
+(4000,4000,2000) (8000,4000,0) (12000,4000,2000) (16000,4000,200)
+(4000,8000,500) (8000,8000,2000) (12000,8000,200) (16000,8000,2000)
+(4000,12000,2000) (8000,12000,0) (12000,12000,2000) (16000,12000,200)
