+++ resources/demo4.mod1
@@ -1,0 +1,1 @@
+(10000,10000,2000) (5000,5000,5000) (5000,15000,5000) (15000,5000,5000) (15000,15000,5000)
