+++ resources/demo5.mod1
@@ -1,0 +1,1 @@
+(5000,5000,5000) (5000,15000,5000) (15000,15000,4000) (15000,5000,5000) (7400,10000,3000) (7400,15000,3000) (12600,15000,3000) (12600,10000,3000) (5000,10000,5000) (15000,10000,5000) (10000,15000,2000) (10000,12600,1000) (10000,10000,2000) (10000,7400,0) (6200,10000,0) (13800,10000,0) (5000,7400,5000) (15000,7400,5000) (5000,12600,5000) (15000,12600,4500) (10000,5000,5000)
