+++ srcs/gui/mainwindow.hpp
