+++ srcs/main.cpp
@@ -1,0 +1,15 @@
+#include "gui/mainwindow.hpp"
+
+#include <iostream>
+#include <QApplication>
+
+int main(int argc, char *argv[])
+{
+    QApplication a(argc, argv);
+    MainWindow w;
+
+    std::cout << "Hello cutie!" << std::endl;
+
+    w.show();
+    return a.exec();
+}
